Make MLUInt -> (Mg)ModUInt conversions slightly more strict

+++ test/modulo_int.test.jl
@@ -42,7 +42,7 @@ end
     mod_tp = ModUInt{mp_tp, m}
 
     @test convert(mod_tp, convert(mp_tp, 1)) == mod_tp(1)
-    @test convert(mod_tp, convert(MLUInt{3, UInt16}, 1)) == mod_tp(1)
+    @test convert(mod_tp, convert(MLUInt{3, UInt8}, 1)) == mod_tp(1)
     @test convert(mod_tp, convert(mp_tp, 1)) == mod_tp(1)
     @test convert(Int, mod_tp(1)) == 1
 end
